First commit of the Bar Graph GUI widget.
This widget creates a standard bar graph with one data group, or a stacked one with more than one group.
Optionally, it also lets you define a number of status "flags" which appear as blocks underneath the X axis of the graph.
The graph is written to gracefully degrade into an unordered list when styles are turned off.

The template code for the widget requires you specify:
1. The total height of the graph (in pixels).
2. The names and styles of the groups (a number of colours and styles are provided).
3. The names and colours of any flags appearing below the graph.
4. A data provider channel method that will return the data for the graph.

The data provider requires you specify as data:
1. The data point values inside each row (one row becomes one bar on the graph).
2. The values of any flags (boolean means set/unset, number/string will show the flag with that value).
3. Optionally, X axis and "top of bar" labels.
4. Optionally, a supplementary CSS class which applies to the whole bar (requires definition outside of the widget).


git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@16 1b08cf3a-d8dc-418b-9366-0d0b919357ef