Minor change in SquizSuite connection sequence.
When two systems are linked (e.g. update system has many products registered), when the connect request arrives to the destined product, it may or may not have the systemid registered. For example, in case of Update system, it will already know the systemid of the product which is sending the request as Update system is the one generates the systemid.

So let's allow to accept the request even the systemid is already registered. (Previously it refused as it thinks they are already connected.)

Now when we do so, make sure the systemid found does not have URL or the certificate containing the public key yet to prevent third party from pretending themselves to be one.

If this doesn't make sense, never mind. this is the changes part of Update system implementation.




git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@192 1b08cf3a-d8dc-418b-9366-0d0b919357ef