Changes in suiteConnectAck message processing.
Once the initial request is made to the other system for the suite connection, the system is waiting for suiteConnectAck message to come back meaning an approval.

When the other system approves the initial request, the product entry with a temporary systemid needs to be clean up as the system doesn't know anything about the system it requested. Hence, populate the database with new entry with full connection details and send the last acknowledgement back.



git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@353 1b08cf3a-d8dc-418b-9366-0d0b919357ef