When SquizSuite systems communicate each other (using encryption), then requested system, action name and return type also needs to be encrypted. Previously, only the arguments to the action call were included in the encrypted message and the others were transmitted unencrypted.
Fix made to API and SquizSuite system that the message decryption will happen before API system looks up the system and action name from the request. If it's not found in $_POST var, then it looks up $_POST['_enc'] fields to get the values.

As a result, every details of API request are encrypted.



git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@355 1b08cf3a-d8dc-418b-9366-0d0b919357ef