Add new setting "background" to Pie graph.
Background is a style applied to the background of the Pie graph, mainly to compensate for Webkit's inability to give SVGs in OBJECT tags a transparent background. (It makes it transparent when using an IMG tag, but Firefox isn't likely to support this until 4.0.)

Background styles committed are white (default), grey, transparent. If further styles need to be defined, they should be done from the widget's CSS file (as .GUIGraphPie.background-* styles).


git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@598 1b08cf3a-d8dc-418b-9366-0d0b919357ef