Fixed initial date setting on the Date Field widget.
The initial commit wouldn't set the initial date properly. This was initially due to string !== int issues, however it has resulted to changes to the day + month Select boxes.

Originally an attempt was made to make the Select option keys the padded day or month name - for ease of handling in JavaScript when constructing the date. However it made it impossible to set an initial date of October or later, or 10th of the month or later - although a key of '01' to '09' remains a string, '10' to '31' becomes an int, and we cannot match both at the same time with a '==='.

Change to using integer keys throughout and handle single-digit day and month appropriately in the JavaScript.


git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@605 1b08cf3a-d8dc-418b-9366-0d0b919357ef