wsDownloadUpdate() action has been changed.
Instead of sending SquizUpdate::updateService() action to do the download, now it requests to SquizUpdate::downloadPatchFile() to get the direct file download.

The incoming file is encrypted the product's public key, so the result needs to be decrypted. Openssl decryption code has been manually added to improve the performance in handling a very large size of patch file.



git-svn-id: http://public-cvs.squiz.net/svn/squiz-framework@841 1b08cf3a-d8dc-418b-9366-0d0b919357ef